re-base taxonomy tool - validate
