+++ cmd/taxonomy_verify.go
@@ -1,0 +1,47 @@
+// Copyright 2020 IBM Corp.
+// SPDX-License-Identifier: Apache-2.0
+package cmd
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/mesh-for-data/mesh-for-data/pkg/taxonomy"
+	"github.com/spf13/cobra"
+	"helm.sh/helm/v3/cmd/helm/require"
+)
+
+// taxonomyCmd represents "schema" command option as part of the "taxonomy" commands
+var schemaCmd = &cobra.Command{
+	Use:   "schema",
+	Short: "commands for working with taxonomy JSON schema files",
+	Args:  require.NoArgs,
+}
+
+// validateCmd  "validate" schema file command under the schema command
+var validateCmd = &cobra.Command{
+	Use:   "validate FILE",
+	Short: "validates a taxonomy JSON schema file",
+	Args:  require.ExactArgs(1),
+	ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		if len(args) == 0 {
+			// Allow file completion when completing the argument for the name
+			// which could be a path
+			return nil, cobra.ShellCompDirectiveDefault
+		}
+		// No more completions, so disable file completion
+		return nil, cobra.ShellCompDirectiveNoFileComp
+	},
+	RunE: func(cmd *cobra.Command, args []string) error {
+		err := taxonomy.ValidateSchema(args[0])
+		if err == nil {
+			fmt.Println("Validate: " + strings.Join(args, " ") + " successfully")
+		}
+		return err
+	},
+}
+
+func init() {
+	taxonomyCmd.AddCommand(schemaCmd)
+	schemaCmd.AddCommand(validateCmd)
+}
